Changed GET_HEAD response, added HEAD message plus handler
Changed the GET_HEAD reponse to send back the hash of the head in
a HEAD message, that way we can distinguish between a response for
the head and a normal block coming in. Added HEAD format and the
handler for it which responds by requesting blocks starting from
the recieved head if that is not in our blockchain.