Update Node to maintain blockchain, miner based off node
Changed the Node class to maintain its own blockchain while it is
connected to the network. Performs all actions such as rebroadcasting
transactions, add to blockchain, etc. In the old sense, it is now a
miner that does not mine. Updated the Miner class to be based off of
the new iteration of the Node class. Also update version number of
Node class