Make server a node (#115)
* Modify Handling to allow not mining

Updated the MessageHandler and HandleMessageThread to based on a
boolean flag, perform mining operations or not.

* Update Node to maintain blockchain, miner based off node

Changed the Node class to maintain its own blockchain while it is
connected to the network. Performs all actions such as rebroadcasting
transactions, add to blockchain, etc. In the old sense, it is now a
miner that does not mine. Updated the Miner class to be based off of
the new iteration of the Node class. Also update version number of
Node class

* Main allows creation of Node or Miner from args

* Make Server a Node

* Update error message

* Fix creating node or miner